+++ pkg/common/config/config.go
@@ -2,7 +2,9 @@ package config
 
 import (
 	"encoding/json"
+	"errors"
 	"io/ioutil"
+	"log"
 )
 
 type AdminConfig struct {
@@ -30,9 +32,29 @@ type TwitchAuthConfig struct {
 }
 
 type TwitchWebhookConfig struct {
-	HostPrefix       string
-	Secret           string
-	LeaseTimeSeconds int
+	Secret string
+
+	// HostPrefix is deprecated since 2021-07-11
+	HostPrefix *string
+
+	// LeaseTimeSeconds is deprecated since 2021-07-11
+	LeaseTimeSeconds *int
+}
+
+func (c *TwitchWebhookConfig) Validate() error {
+	if len(c.Secret) < 10 || len(c.Secret) > 100 {
+		return errors.New("twitch WebHook secret must be at least 10 characters and at most 100 characters")
+	}
+
+	if c.HostPrefix != nil {
+		log.Println("Twitch.Webhook.HostPrefix is deprecated - we now rely on Web.Domain + Web.Secure to figure out the webhook callback URL.")
+	}
+
+	if c.LeaseTimeSeconds != nil {
+		log.Println("Twitch.Webhook.LeaseTimeSeconds is deprecated. This value serves no purpose for Twitch's EventSub.")
+	}
+
+	return nil
 }
 
 type AuthTwitchConfig struct {
@@ -99,13 +121,6 @@ var defaultConfig = Config{
 	PostgreSQL: PostgreSQLConfig{
 		DSN: "host=/var/run/postgresql database=pajbot2 sslmode=disable",
 	},
-	Auth: authConfig{
-		Twitch: AuthTwitchConfig{
-			Webhook: TwitchWebhookConfig{
-				LeaseTimeSeconds: 24 * 3600,
-			},
-		},
-	},
 }
 
 /*
@@ -124,4 +139,12 @@ func LoadConfig(path string) (*Config, error) {
 	}
 
 	return &config, nil
+}
+
+func (c *Config) Validate() error {
+	if err := c.Auth.Twitch.Webhook.Validate(); err != nil {
+		return err
+	}
+
+	return nil
 }