+++ pkg/twitch/bot.go
@@ -16,7 +16,6 @@ import (
 
 	twitch "github.com/gempir/go-twitch-irc/v2"
 	"github.com/pajbot/pajbot2/pkg"
-	"github.com/pajbot/pajbot2/pkg/apirequest"
 	"github.com/pajbot/pajbot2/pkg/channels"
 	"github.com/pajbot/pajbot2/pkg/common"
 	"github.com/pajbot/pajbot2/pkg/users"
@@ -618,14 +617,6 @@ func (b *Bot) HandleMessage(channelName string, user twitch.User, rawMessage *tw
 }
 
 func (b *Bot) HandleRoomstateMessage(message twitch.RoomStateMessage) {
-	if len(message.Tags) > 2 {
-		if channelID, ok := message.Tags["room-id"]; ok {
-			// Joined channel
-			b.streamStore.JoinStream(&SimpleAccount{channelID, message.Channel})
-		} else {
-			fmt.Println("room-id not set in roomstate message:", message.Raw)
-		}
-	}
 	subMode := ModeUnset
 
 	channel := &channels.TwitchChannel{
@@ -670,34 +661,34 @@ func (b *Bot) HandleClearChatMessage(message *twitch.ClearChatMessage) {
 }
 
 func (b *Bot) StartChatterPoller() {
-	b.ticker = time.NewTicker(5 * time.Minute)
+	// b.ticker = time.NewTicker(5 * time.Minute)
 	// defer close ticker lol
 
-	go func() {
-		for range b.ticker.C {
-			chatters, err := apirequest.TwitchWrapper.API().TMI().GetChatters("pajlada")
-			if err != nil {
-				continue
-			}
+	// go func() {
+	// 	for range b.ticker.C {
+	// 		chatters, err := apirequest.TwitchWrapper.API().TMI().GetChatters("pajlada")
+	// 		if err != nil {
+	// 			continue
+	// 		}
 
-			var usernames []string
-			usernames = append(usernames, chatters.Moderators...)
-			usernames = append(usernames, chatters.Staff...)
-			usernames = append(usernames, chatters.Admins...)
-			usernames = append(usernames, chatters.GlobalMods...)
-			usernames = append(usernames, chatters.Viewers...)
-			userIDs := b.GetUserStore().GetIDs(usernames)
+	// 		var usernames []string
+	// 		usernames = append(usernames, chatters.Moderators...)
+	// 		usernames = append(usernames, chatters.Staff...)
+	// 		usernames = append(usernames, chatters.Admins...)
+	// 		usernames = append(usernames, chatters.GlobalMods...)
+	// 		usernames = append(usernames, chatters.Viewers...)
+	// 		userIDs := b.GetUserStore().GetIDs(usernames)
 
-			userIDsSlice := make([]string, len(userIDs))
-			i := 0
-			for _, userID := range userIDs {
-				userIDsSlice[i] = userID
-				i++
-			}
+	// 		userIDsSlice := make([]string, len(userIDs))
+	// 		i := 0
+	// 		for _, userID := range userIDs {
+	// 			userIDsSlice[i] = userID
+	// 			i++
+	// 		}
 
-			b.BulkEdit("pajlada", userIDsSlice, 25)
-		}
-	}()
+	// 		b.BulkEdit("pajlada", userIDsSlice, 25)
+	// 	}
+	// }()
 }
 
 type PointServer struct {
@@ -960,8 +951,6 @@ func (b *Bot) JoinChannel(channelID string) error {
 
 	botChannel := &BotChannel{
 		ID: id,
-
-		stream: b.streamStore.GetStream(&botChannelUser),
 
 		channel: botChannelUser,
 	}