+++ pkg/web/controller/api/auth/twitch/bot.go
@@ -3,7 +3,7 @@ package twitch
 import (
 	"net/http"
 
-	"github.com/dankeroni/gotwitch/v2"
+	"github.com/nicklaw5/helix"
 	"github.com/pajbot/pajbot2/pkg/web/state"
 	"github.com/pajbot/utils"
 	"golang.org/x/oauth2"
@@ -11,7 +11,7 @@ import (
 
 func onBotAuthenticated(
 	w http.ResponseWriter, r *http.Request,
-	self gotwitch.ValidateResponse, oauth2Token *oauth2.Token, stateData *stateData) {
+	self *helix.ValidateTokenResponse, oauth2Token *oauth2.Token, stateData *stateData) {
 	const queryF = `
 INSERT INTO bot (twitch_userid, twitch_username, twitch_access_token, twitch_refresh_token, twitch_access_token_expiry)
     VALUES ($1, $2, $3, $4, $5) ON CONFLICT (twitch_userid)
@@ -23,7 +23,7 @@ INSERT INTO bot (twitch_userid, twitch_username, twitch_access_token, twitch_ref
         twitch_refresh_token = $4,
         twitch_access_token_expiry = $5`
 	c := state.Context(w, r)
-	_, err := c.SQL.Exec(queryF, self.UserID, self.Login, oauth2Token.AccessToken, oauth2Token.RefreshToken, oauth2Token.Expiry) // GOOD
+	_, err := c.SQL.Exec(queryF, self.Data.UserID, self.Data.Login, oauth2Token.AccessToken, oauth2Token.RefreshToken, oauth2Token.Expiry) // GOOD
 	if err != nil {
 		_ = utils.WebWriteError(w, 500, "Error inserting bot, admin should check console logs NaM")
 		return