+++ pkg/web/controller/api/auth/twitch/streamer.go
@@ -3,11 +3,11 @@ package twitch
 import (
 	"net/http"
 
-	"github.com/dankeroni/gotwitch/v2"
+	"github.com/nicklaw5/helix"
 	"golang.org/x/oauth2"
 )
 
 func onStreamerAuthenticated(
 	w http.ResponseWriter, r *http.Request,
-	self gotwitch.ValidateResponse, oauth2Token *oauth2.Token, stateData *stateData) {
+	self *helix.ValidateTokenResponse, oauth2Token *oauth2.Token, stateData *stateData) {
 }